+++ app/controllers/notes_to_self_controller.rb
@@ -1,7 +1,8 @@
 # people can write notes to self to remind them why they added a show
 class NotesToSelfController < ApplicationController
   def edit
-    authorize! { current_human.present? }
+    require_authentication!
+    authorize! { true }
 
     @show = Show.find_by!(slug: params.expect(:show_slug))
     @my_show = MyShow.find_by!(human: current_human, show: @show)